Various updates for istio demo scripts. Now works with GCE

+++ istio/create-nginx-examples.sh
@@ -11,8 +11,16 @@ function create_certs() {
   TYPE="$1"
   CERT_DOMAIN="$2"
   NAMESPACE="$3"
-  test -f ${CERT_DIR}/${TYPE}.${NAMESPACE}.key ||  openssl req -out ${CERT_DIR}/${TYPE}.${NAMESPACE}.csr -newkey rsa:2048 -nodes -keyout ${CERT_DIR}/${TYPE}.${NAMESPACE}.key -subj "/CN=${CERT_DOMAIN}/O=RedHat"
+  test -f ${CERT_DIR}/${TYPE}.${NAMESPACE}.key || openssl req -out ${CERT_DIR}/${TYPE}.${NAMESPACE}.csr -newkey rsa:2048 -nodes -keyout ${CERT_DIR}/${TYPE}.${NAMESPACE}.key -subj "/CN=${CERT_DOMAIN}/O=RedHat"
+  if [[ $? -ne 0 ]]; then
+    echo "ERROR: Cert generation for $CERT_DOMAIN failed!"
+    exit 1
+  fi
   test -f ${CERT_DIR}/${TYPE}.${NAMESPACE}.crt || openssl x509 -req -sha256 -days 365 -CA ${CERT_DIR}/ca.crt -CAkey ${CERT_DIR}/ca.key -set_serial 0 -in ${CERT_DIR}/${TYPE}.${NAMESPACE}.csr -out ${CERT_DIR}/${TYPE}.${NAMESPACE}.crt
+  if [[ $? -ne 0 ]]; then
+    echo "ERROR: Cert generation for $CERT_DOMAIN failed!"
+    exit 1
+  fi
   oc delete -n $NAMESPACE secret ${TYPE}-credential 2>/dev/null
   oc create -n $NAMESPACE secret tls ${TYPE}-credential --key=${CERT_DIR}/${TYPE}.${NAMESPACE}.key --cert=${CERT_DIR}/${TYPE}.${NAMESPACE}.crt
 }
@@ -29,12 +37,12 @@ test -f ${CERT_DIR}/ca.crt || openssl req -x509 -sha256 -nodes -days 365 -newkey
 
 # Istio API Certs
 create_certs edge "edge.${ISTIO_DOMAIN}" istio-system
-create_certs reencrypt "reencrypt.${ISTIO_DOMAIN}" istio-system
+create_certs re "re.${ISTIO_DOMAIN}" istio-system
 create_certs pass "pass.${ISTIO_DOMAIN}" istioapi
 
 # Gateway API Certs
 create_certs edge "edge.${GWAPI_DOMAIN}" gwapi
-create_certs reencrypt "reencrypt.${GWAPI_DOMAIN}" gwapi
+create_certs re "re.${GWAPI_DOMAIN}" gwapi
 create_certs pass "pass.${GWAPI_DOMAIN}" gwapi
 
 # Install gateway api
@@ -43,10 +51,13 @@ kubectl get crd gateways.gateway.networking.k8s.io &> /dev/null || { kubectl kus
 # Configure istioapi examples via istio api
 cat ./nginx-istioapi.yaml | envsubst | oc apply -f -
 cat ./nginx-gwapi.yaml | envsubst | oc apply -f -
+cat ./echo-service-sleeper-istioapi.yaml | envsubst | oc apply -f -
 
 TIMEOUT=60
-while [[ "$GWAPI_LOADBALANCER" == "" ]]; do
-  GWAPI_LOADBALANCER=$(oc -n gwapi get service gateway -o jsonpath='{.status.loadBalancer.ingress[0].hostname}')
+while [[ "$GWAPI_LOADBALANCER_DOMAIN" == "" ]] && [[ "$GWAPI_LOADBALANCER_IP" == "" ]]; do
+  # For AWS, it uses hostname, but for GCE, it uses IP
+  GWAPI_LOADBALANCER_DOMAIN=$(oc -n gwapi get service gateway -o jsonpath='{.status.loadBalancer.ingress[0].hostname}')
+  GWAPI_LOADBALANCER_IP=$(oc -n gwapi get service gateway -o jsonpath='{.status.loadBalancer.ingress[0].ip}')
   echo "Waiting for gateway api loadbalancer to get domain name"
   if [[ "$TIMEOUT" -lt 0 ]]; then
     echo "ERROR: Gateway API loadbalancer never got domain name"
@@ -55,6 +66,14 @@ while [[ "$GWAPI_LOADBALANCER" == "" ]]; do
   TIMEOUT=$((TIMEOUT-1))
   sleep 1
 done
+
+if [[ "$GWAPI_LOADBALANCER_DOMAIN" != "" ]]; then
+  GWAPI_LOADBALANCER="$GWAPI_LOADBALANCER_DOMAIN"
+  DNS_RECORD_TYPE="CNAME"
+else
+  GWAPI_LOADBALANCER="$GWAPI_LOADBALANCER_IP"
+  DNS_RECORD_TYPE="A"
+fi
 
 # Add wildcard to send dns requests to istio via gwapi
 oc apply -f - <<EOF
@@ -68,7 +87,7 @@ metadata:
 spec:
   dnsName: "*.${GWAPI_DOMAIN}."
   recordTTL: 30
-  recordType: CNAME
+  recordType: ${DNS_RECORD_TYPE}
   targets:
   - ${GWAPI_LOADBALANCER}
 EOF
@@ -77,11 +96,11 @@ EOF
 echo "ISITIO API:"
 echo "curl -I http://http.${ISTIO_DOMAIN}"
 echo "curl --cacert ${CERT_DIR}/ca.crt -I https://edge.${ISTIO_DOMAIN}"
-echo "curl --cacert ${CERT_DIR}/ca.crt -I https://reencrypt.${ISTIO_DOMAIN}"
+echo "curl --cacert ${CERT_DIR}/ca.crt -I https://re.${ISTIO_DOMAIN}"
 echo "curl --cacert ${CERT_DIR}/ca.crt -I https://pass.${ISTIO_DOMAIN}"
 echo
 echo "GWAPI:"
 echo "curl -I http://http.${GWAPI_DOMAIN}"
 echo "curl --cacert ${CERT_DIR}/ca.crt -I https://edge.${GWAPI_DOMAIN}"
-echo "curl --cacert ${CERT_DIR}/ca.crt -I https://reencrypt.${GWAPI_DOMAIN}"
+echo "curl --cacert ${CERT_DIR}/ca.crt -I https://re.${GWAPI_DOMAIN}"
 echo "curl --cacert ${CERT_DIR}/ca.crt -I https://pass.${GWAPI_DOMAIN}"
